+++ src/runtime/components/EditorMentionMenu.vue
@@ -78,7 +78,10 @@ onMounted(async () => {
         .deleteRange(range)
         .insertContent({
           type: 'mention',
-          attrs: item
+          attrs: {
+            ...item,
+            mentionSuggestionChar: props.char
+          }
         })
         .run()
     },